New options for greater control on mounting

+++ main.cc
@@ -28,6 +28,8 @@ struct Options
     bool debug;
     bool mount_proc;
     bool mount_sys;
+    vector<string> extra_mounts;
+    bool mount_program;
 
     Options()
     {
@@ -37,12 +39,14 @@ struct Options
         debug = false;
         mount_proc = false;
         mount_sys = false;
+        mount_program = true;
     }
 
     Options(const Options& o)
      : timeout_ms{o.timeout_ms},
        uid{o.uid}, gid{o.gid}, debug{o.debug},
-       mount_proc{o.mount_proc}, mount_sys{o.mount_sys}
+       mount_proc{o.mount_proc}, mount_sys{o.mount_sys},
+       extra_mounts{o.extra_mounts}, mount_program{o.mount_program}
     {
     }
 
@@ -62,6 +66,13 @@ void Options::Usage(const char* prog)
     cerr << "    -g gid     Run command with group gid\n";
     cerr << "    -p         Mount /proc\n";
     cerr << "    -s         Mount /sys\n";
+    cerr << "    -m path    Mount path under /mnt/`basename path`\n";
+    cerr << "    -M         Do not mount program\n";
+    cerr << "\n";
+    cerr << "The -m option can be repeated to mount multiple paths.\n";
+    cerr << "If the -M option is not specified, the program is mounted at\n";
+    cerr << "/program which is useful for programs that are not installed\n";
+    cerr << "in standard locations such as /bin or /usr/bin\n";
     cerr << "\n";
 }
 
@@ -69,7 +80,7 @@ Options Options::Parse(int& argc, char**& argv)
 {
     int opt;
     Options options;
-    while ((opt = getopt(argc, argv, "+dt:u:g:ps")) != -1) {
+    while ((opt = getopt(argc, argv, "+dt:u:g:psm:M")) != -1) {
         switch (opt) {
             case 'd':   options.debug = true;   break;
             case 't':
@@ -89,6 +100,8 @@ Options Options::Parse(int& argc, char**& argv)
             case 'g':   options.gid = atoi(optarg);     break;
             case 'p':   options.mount_proc = true;      break;
             case 's':   options.mount_sys = true;       break;
+            case 'm':   options.extra_mounts.push_back(optarg); break;
+            case 'M':   options.mount_program = false;  break;
             default:
             {
                 Usage(argv[0]);
@@ -111,6 +124,12 @@ void Options::Log()
     log << "  Debug: " << debug << "\n";
     log << "  Mount /proc: " << mount_proc << "\n";
     log << "  Mount /sys: " << mount_sys << "\n";
+    log << "  Extra mounts (" << extra_mounts.size() << "):\n";
+    for (auto& x : extra_mounts)
+    {
+        log << "    " << x << "\n";
+    }
+    log << "  Mount program: " << mount_program << "\n";
 }
 
 class Sandbox
@@ -123,7 +142,7 @@ class Sandbox
         ChangeMode(rootfs, 0755);
         log << " rootfs = " << rootfs << "\n";
         CreatePrivateMount(rootfs);
-        for (auto& folder : to_mount)
+        for (auto& folder : always_mount)
         {
             if (PathExists(folder))
             {
@@ -132,10 +151,30 @@ class Sandbox
                 CreateFolder(mount_point);
             }
         }
-        program_mount_point = rootfs + program_path;
-        log << " Creating program mount point " << program_mount_point << "\n";
-        ofstream pfs(program_mount_point);
-        pfs.close();
+        log << " Creating folder " << rootfs + "/mnt" << "\n";
+        CreateFolder(rootfs + "/mnt");
+        for (auto& path : options.extra_mounts)
+        {
+            string mount_point = rootfs + "/mnt/" + BaseName(path);
+            if (IsDirectory(path))
+            {
+                log << " Creating folder " << mount_point << "\n";
+                CreateFolder(mount_point);
+            }
+            if (IsRegularFile(path))
+            {
+                log << " Creating file " << mount_point << "\n";
+                ofstream pfs(mount_point);
+                pfs.close();
+            }
+        }
+        if (options.mount_program)
+        {
+            program_mount_point = rootfs + program_path;
+            log << " Creating program mount point " << program_mount_point << "\n";
+            ofstream pfs(program_mount_point);
+            pfs.close();
+        }
     }
 
     ~Sandbox()
@@ -145,9 +184,27 @@ class Sandbox
             try { // We don't want to throw any exceptions from a dtor
                 log << "\n[" << getpid() << "] ~Sandbox():\n";
                 log << " ctor_pid = " << ctor_pid << "\n";
-                log << " Deleting " << program_mount_point << "\n";
-                DeleteFile(program_mount_point);
-                for (auto& folder : to_mount)
+                if (options.mount_program)
+                {
+                    log << " Deleting " << program_mount_point << "\n";
+                    DeleteFile(program_mount_point);
+                }
+                for (auto& path : options.extra_mounts)
+                {
+                    string mount_point = rootfs + "/mnt/" + BaseName(path);
+                    log << " Deleting " << mount_point << "\n";
+                    if (IsDirectory(mount_point))
+                    {
+                        DeleteFolder(mount_point);
+                    }
+                    if (IsRegularFile(mount_point))
+                    {
+                        DeleteFile(mount_point);
+                    }
+                }
+                log << " Deleting " << rootfs + "/mnt" << "\n";
+                DeleteFolder(rootfs + "/mnt");
+                for (auto& folder : always_mount)
                 {
                     string mount_point = rootfs + folder;
                     if (PathExists(mount_point))
@@ -174,7 +231,7 @@ class Sandbox
     }
 
   private:
-    static vector<string> to_mount;
+    static vector<string> always_mount;
     Options options;
     string rootfs;
     pid_t ctor_pid;
@@ -191,7 +248,7 @@ class Sandbox
             Unshare(flags);
             MarkMountPointPrivate("/");
 
-            for (auto& folder : to_mount)
+            for (auto& folder : always_mount)
             {
                 if (PathExists(folder))
                 {
@@ -201,17 +258,43 @@ class Sandbox
                 }
             }
 
-            log << " Mounting program " << args[0] << " at " << program_mount_point << "\n";
-            BindMount(args[0], program_mount_point);
-            args[0] = strdup(program_path);
+            for (auto& path : options.extra_mounts)
+            {
+                string mount_point = rootfs + "/mnt/" + BaseName(path);
+                if (PathExists(mount_point))
+                {
+                    log << " Mounting " << path << " at " << mount_point << "\n";
+                    BindMount(path, mount_point);
+                }
+            }
+
+            if (options.mount_program)
+            {
+                log << " Mounting program " << args[0] << " at " << program_mount_point << "\n";
+                BindMount(args[0], program_mount_point);
+                args[0] = strdup(program_path);
+            }
 
             ForkCallWait([&]() { chroot_run(args); });
 
             log << "\n Unmounting...\n";
-            log << " Unmounting " << program_mount_point << "\n";
-            Unmount(program_mount_point);
+            if (options.mount_program)
+            {
+                log << " Unmounting " << program_mount_point << "\n";
+                Unmount(program_mount_point);
+            }
 
-            for (auto& folder : to_mount)
+            for (auto& path : options.extra_mounts)
+            {
+                string mount_point = rootfs + "/mnt/" + BaseName(path);
+                if (PathExists(mount_point))
+                {
+                    log << " Unmounting " << mount_point << "\n";
+                    Unmount(mount_point);
+                }
+            }
+
+            for (auto& folder : always_mount)
             {
                 string mount_point = rootfs + folder;
                 if (PathExists(mount_point))
@@ -301,7 +384,7 @@ class Sandbox
     }
 };
 
-vector<string> Sandbox::to_mount{ "/bin", "/etc", "/lib", "/lib32", "/lib64", "/usr" };
+vector<string> Sandbox::always_mount{ "/bin", "/etc", "/lib", "/lib32", "/lib64", "/usr" };
 
 int main(int argc, char* argv[])
 {
